refactor: new Expression and Statment facade
Move span information for expressions and statements into a struct for
all types. The information for each node type is still stored in the
enum. This removes all tokens from the tree and makes it easier to work
with. Gives more information about positioning in a consistent way which
can be used to improve diagnostics.